Rename remaining files
