+++ src/pathpicker/char_code_mapping.py
