+++ src/pathpicker/color_printer.py
