+++ src/tests/test_key_bindings_parsing.py
@@ -6,9 +6,8 @@ import sys
 
 from tests.key_bindings_for_test import KEY_BINDINGS_FOR_TEST_CONFIG_CONTENT
 from tests.key_bindings_for_test import KeyBindingsForTest
-from pathpicker.keyBindings import KeyBindings
+from pathpicker.key_bindings import KeyBindings
 import tempfile
-from pathpicker import keyBindings
 import unittest
