+++ src/tests/test_parsing.py
@@ -8,7 +8,7 @@ import os
 
 from pathpicker import parse
 from tests.local_test_cases import LOCAL_TEST_CASES
-from pathpicker.formattedText import FormattedText
+from pathpicker.formatted_text import FormattedText
 from pathpicker import format
 
 fileTestCases = [
